Make JLine/ReplxxCli.scala and CliImpl.scala separate
